Source code documentation

Contents

Source code documentation

Contents

SeasonalEpid structs

Time parameters

Biological parameters

States

SeasonalEpid functions

Functions in functions.jl

Reproducing figures

SeasonalEpid structs

NOTE: If you want to know more about what a struct contains, type fieldnames(structname).

Time parameters

SeasonalEpid.TimeParam — Type

Stock all time variables.

Biological parameters

These structs can be called biological parameters but note that they contain also some informations about the model, like if it is elaborate, and the number of states.

SeasonalEpid.Param — Type

This is the most global param. Subtypes of Param are also abstract type.

SeasonalEpid.Elaborate1Strain — Type

This param is also a global param but it is only for elaborate 1 strain models.

SeasonalEpid.Elaborate2Strains — Type

This param is also a global param but it is only for elaborate 2 strain2 models.

SeasonalEpid.Compact1Strain — Type

This param is also a global param but it is only for compact 1 strain models.

SeasonalEpid.Compact2Strains — Type

This param is also a global param but it is only for compact 2 strains models.

SeasonalEpid.ParamSoilborneCompact1Strain — Type

This param is very specific and it is built to construct soilborne compact 1 strain models.

SeasonalEpid.ParamAirborneElaborate1Strain — Type

This param is very specific and it is built to construct airborne elaborate 1 strain models.

States

SeasonalEpid.StateParam0 — Type

This is the most global Stateparam. Subtypes of StateParam0 are also abstract type.

SeasonalEpid.StateCompact — Type

This stateparam is very specific to compact 1 strain models and it stocks model's states at the beginning of a season.

SeasonalEpid.StateElaborate — Type

This stateparam is very specific to elaborate 1 strain models and it stocks model's states at the beginning of a season.

SeasonalEpid functions

Functions in functions.jl

SeasonalEpid.GrowingSeason — Method

GrowingSeason(State0::SVector, param::Compact1Strain, t::Real)

This is the function to enter in ODEProblem from DifferentialEquations.jl.

For a compact model, it returns the ODE associated to the growing season.

See also WinterSeason, growing.

SeasonalEpid.GrowingSeason — Method

GrowingSeason(State0::SVector, param::ParamAirborneElaborate1Strain, t::Real)

This is the function to enter in ODEProblem from DifferentialEquations.jl.

For an elaborate model, it returns the ODE associated to the growing season.

See also WinterSeason, growing.

SeasonalEpid.WinterSeason — Method

WinterSeason(State0::SVector, param::Elaborate1Strain, t::Real)

This is the function to enter in ODEProblem from DifferentialEquations.jl.

It only exists for elaborate models, and it returns the ODE associated to the winter season.

See also GrowingSeason, winter.

SeasonalEpid.displaysim — Method

displaysim(nyears::Int64, sp::StateParam0, param::Param; tp::TimeParam=TimeParam())

Make a simulation of n years for any model. Plot the solutions of this simulation.

See also simule.

SeasonalEpid.fill_mat — Method

fill_mat(nyears::Int64,	sp::StateParam0, param::Param; tp::TimeParam=TimeParam())

Construct an empty named matrix to stock nyears of simulation.

Labels are filled automatically.

See also simule.

SeasonalEpid.growing — Method

growing(sp::StateParam0, param::Param; tp::TimeParam=TimeParam())

Simulates the growing season for any model, using ODEProblem from DifferentialEquations.jl.

It returns a vector of vectors that contains the simulation for a season, and also the last values of the simulation.

See also GrowingSeason, winter, simule.

SeasonalEpid.isWinter — Method

isWinter(t,tp)

From a time vector, returns a new vector of 0 and 1 to identify growing and winter seasons.

SeasonalEpid.simule — Method

simule(nyears::Int64, sp::StateParam0, param::Param; tp::TimeParam=TimeParam())

Simulate n years for any model.

It returns a named matrix that contains n years of simulation.

SeasonalEpid.simule — Method

simule(sp::StateParam0,	param::Param; tp::TimeParam=TimeParam())

Simulate a year for any model.

It returns a vector of vectors that contains one year of simulation, and also the last values of the simulation.

See also growing, winter, yeartransition, fill_mat, displaysim.

SeasonalEpid.winter — Method

winter(res_end,	param::Elaborate1Strain; tp::TimeParam=TimeParam())

Compute new initial conditions from the last values of the previous growing season. Then it simulates the winter season for elaborate 1 strain models, using ODEProblem from DifferentialEquations.jl.

It returns a vector of vectors that contains the simulation, and the last values of the simulation.

See also WinterSeason, winter, simule.

SeasonalEpid.yeartransition — Method

yeartransition(res_end,	param::ParamAirborneElaborate1Strain; tp::TimeParam=TimeParam())

Compute new initial conditions from the last values of winter season simulation.

It returns a StateElaborate object.

See also simule.

SeasonalEpid.yeartransition — Method

yeartransition(res_end,	param::ParamSoilborneCompact1Strain; tp::TimeParam=TimeParam())

Compute new initial conditions from the last values of growing season simulation.

It returns a StateCompact object.

See also simule.

Reproducing figures

Maybe we can show here few examples.